Post finished

+++ controllers/postController.js
@@ -4,21 +4,32 @@ const Comment = require('../Models/post').Comment;
 const Reply = require('../Models/post').Reply;
 const Share = require('../Models/post').Share;
 const User = require('../Models/user');
-const notificationService=require('../Services/notification')
 const upload = require("../middlewares/uploadFile");
+const path = require("path");
+const fs = require("fs");
 
+//get all posts
+const getAllPosts = async (req, res) => {
+    try {
+      const userId =req.userId;
+      const posts = await Post.find({author:userId}).populate('author');
+      res.status(200).json(posts);
+    } catch (error) {
+      console.error(error);
+      res.status(500).json({ error: 'Internal Server Error' });
+    }
+};
 //create post
 const addPost = async (req, res, next) => {
     try{
         const files = req.files;
-        const author=req.userId;
+        const author = req.userId;
         const { content } = req.body;
 
         // Create a new post instance
         const newPost = new Post({
             author: author, 
             content: content,
-            dateCreated : Date.now(),
             attachments: files.map(file => ({ 
                 type: file.mimetype.split('/')[0], 
                 url: file.path // Store the file path as the URL (you may need to adjust this based on your file storage setup)
@@ -35,17 +46,14 @@ const addPost = async (req, res, next) => {
         }
         next(err);
     }
-}
+};
 
 //update post
 const updatePost = async (req, res, next) => {
     try{
-        const {content}=req.body;
-        
         //find the id of the post to be updated
         const post = await Post.findById(req.params.id);
-        await post.updateOne({$set:{content}});
-
+        await post.updateOne({$set:req.body});
         res.status(200).json({message:'The post has been successfully updated'})
     }
     catch(err){
@@ -55,7 +63,7 @@ const updatePost = async (req, res, next) => {
         }
         next(err);  
     }
-}
+};
 
 //delete post
 const deletePost = async (req, res, next) => {
@@ -72,7 +80,7 @@ const deletePost = async (req, res, next) => {
       }
       next(err);  
     }
-}
+};
 
 //react post
 const reactPost = async (req, res, next) => {
@@ -85,7 +93,7 @@ const reactPost = async (req, res, next) => {
 
       //Check if the user has already reacted to the post
       let userReacted = post.reacts.users.includes(req.userId);
-      if (!userReacted) {
+      if (! userReacted) {
         // User has not reacted, allow reacting to the post
         post.reacts.count += 1;
         post.reacts.users.push(req.userId);
@@ -120,14 +128,12 @@ const addComment  = async (req,res,next)=>{
             return res.status(404).json({ error: 'This post does not exist' });
         }
         const files = req.files;
-        const author=req.userId;
-        const  {content} = req.body;
-
+        const author = req.userId;
+        const  { content } = req.body;
 
         const newComment = new  Comment ({
             author: author,
             content:content,
-            dateCreated : Date.now(),
             attachments: files.map(file => ({
                 type: file.mimetype.split('/')[0], 
                 url: file.path
@@ -136,9 +142,6 @@ const addComment  = async (req,res,next)=>{
         post.comments.push(newComment);
         await post.save(); 
         await newComment.save();
-        TODO:
-        //send notification
-        await notificationService.addNotification(`some one add comment in your post`,[post.author],'comment');
         res.status(200).json({ message: 'Your comment was shared.' });
 
     } catch (err) {
@@ -150,11 +153,40 @@ const addComment  = async (req,res,next)=>{
     }
 };
 
+//update comment
+const updateComment = async (req, res, next) => {
+    try{
+        // Find the post to which the comment belongs
+        const comment = await Comment.findById(req.params.id);
+        await comment.updateOne({$set:req.body});
+        res.status(200).json({message:'The comment has been successfully updated'})
+    }
+    catch(err){
+        console.error('Error updating post.', err);
+        if (!err.statusCode) {
+            err.statusCode = 500;
+        }
+        next(err);  
+    }
+};
+
+//delete comment
 const deleteComment = async (req, res, next) => {
     try{
-        //find the id of the post to be deleted
-        const comment = await Comment.findById(req.params.id);
+        // Find the post to which the comment belongs
+        const post = await Post.findById(req.params.postId);
+        if (!post) {
+            return res.status(404).json({ message: "Post not found" });
+        }
+        //find the id of the comment to be deleted
+        const comment = await Comment.findById(req.params.commentId);
+        if (!comment) {
+            return res.status(404).json({ message: "Comment not found" });
+        }
+        
+        post.comments.pull(comment);
         await comment.deleteOne();
+        await post.save();
         res.status(200).json({message:'Your comment has been successfully deleted.'})
     }
     catch(err){
@@ -164,7 +196,8 @@ const deleteComment = async (req, res, next) => {
       }
       next(err);  
     }
-}
+};
+
 //react for a comment
 const reactComment = async (req, res, next) => {
     try {
@@ -175,17 +208,17 @@ const reactComment = async (req, res, next) => {
         }
 
       //Check if the user has already reacted to the comment
-      let userReacted = comment.reacts.users.includes(req.body.author);
+      let userReacted = comment.reacts.users.includes(req.userId);
       if (! userReacted) {
         // User has not reacted, allow reacting to the comment
         comment.reacts.count += 1;
-        comment.reacts.users.push(req.body.author);
+        comment.reacts.users.push(req.userId);
         userReacted = true;
         res.status(200).json({message:'You liked the comment.'});
       } else {
         // User has already reacted, allow disliking the comment
         comment.reacts.count -= 1;
-        comment.reacts.users = comment.reacts.users.pull(req.body.author);
+        comment.reacts.users = comment.reacts.users.pull(req.userId);
         userReacted = false;
         res.status(200).json({message:'You disliked the comment.'});
       }
@@ -205,20 +238,36 @@ const reactComment = async (req, res, next) => {
 
 //reply for a comment
 const replyComment  = async (req,res,next)=>{
-    try{    
-        const comment = await Comment.findById(req.params.id);
+    try{   
+        // Find the post to which the comment belongs
+        const post = await Post.findById(req.params.postId);
+        if (!post) {
+            return res.status(404).json({ message: "Post not found" });
+        }
+        const comment = await Comment.findById(req.params.commentId);
         if (!comment) {
             return res.status(404).json({ error: 'This comment does not exist' });
         }
+        const  files = req.files;
+        const author = req.userId;
+        const { content } = req.body;
 
-        const  reply = new Reply (req.body);
+        const  reply = new Reply ({
+            author: author,
+            content: content,
+            attachments: files.map(file => ({
+                type: file.mimetype.split('/')[0], 
+                url: file.path
+            }))
+        });
+        comment.replies = comment.replies || []; // Initialize replies if it doesn't exist
 
-        comment.replies = comment.replies || {}; // Initialize replies if it doesn't exist
-        
         comment.replies.push(reply);
+        comment.count += 1;
+        
         await reply.save(); 
         await comment.save();  
-        
+        await post.save();
         res.status(200).json({ message: 'Your reply was shared.' });
 
     } catch (err) {
@@ -230,40 +279,112 @@ const replyComment  = async (req,res,next)=>{
     }
 };
 
-//save post
-const savePosts = async(req, res, next) =>{
-    try{
-        const user = await User.findById(req.body);
-        if (!user) {
-            return res.status(404).json({ error: 'User not found' });
+//react for a reply
+const reactReply = async (req, res, next) => {
+    try {
+        const comment = await Comment.findById(req.params.commentId);
+        if (!comment) {
+            return res.status(404).json({ error: 'This comment does not exist' });
         }
 
-        const post = await Post.findById(req.params.id);
-        if (!post) {
-            return res.status(404).json({ error: 'Post not found' });
+        const reply = await Reply.findById(req.params.replyId);
+        if (!reply) {
+            return res.status(404).json({ error: 'This reply does not exist' });
         }
 
-        //Check if the user has already saved the post
-        const isSaved = User.savedPosts.includes(req.params.id);
-        if (isSaved) {
-            user.savedPosts.push(req.params.id);
-            await user.save();
-            res.status(200).json({ message: 'You saved the post.' });
-        } else{
-            // Remove the post ID from savedPosts to unsave it
-            user.savedPosts = user.savedPosts.filter((postId) => postId !== req.params.id);
-            await user.save();
-            res.status(200).json({ message: 'You unsaved the post.' });
-        }
-        
-    } catch(err){
-        console.error("Error saving the  post.", err);
+      //Check if the user has already reacted to the comment
+      let userReacted = reply.reacts.users.includes(req.userId);
+      if (! userReacted) {
+        // User has not reacted, allow reacting to the comment
+        reply.reacts.count += 1;
+        reply.reacts.users.push(req.userId);
+        userReacted = true;
+        res.status(200).json({message:'You liked the comment.'});
+      } else {
+        // User has already reacted, allow disliking the comment
+        reply.reacts.count -= 1;
+        reply.reacts.users = reply.reacts.users.pull(req.userId);
+        userReacted = false;
+        res.status(200).json({message:'You disliked the comment.'});
+      }
+
+      await reply.save();
+      await comment.save();
+      res.status(200).json({ message: 'Your comment was reacted.' });
+
+    } catch (err) {
+        console.error("Error reacting comment.", err);
         if (!err.statusCode) {
             err.statusCode = 500;
         }
         next(err);
     }
-}; 
+};
+
+//delete reply
+const deleteRely = async (req, res, next) => {
+    try{
+        
+        const comment = await Comment.findById(req.params.commentId);
+        if (!comment) {
+            return res.status(404).json({ message: "Comment not found" });
+        }
+
+        const reply = await Reply.findById(req.params.replyId);
+        if (!reply) {
+            return res.status(404).json({ message: "Reply not found" });
+        }
+        
+        comment.replies.pull(reply);
+        await reply.deleteOne();
+        await comment.save();
+        res.status(200).json({message:'Your comment has been successfully deleted.'})
+    }
+    catch(err){
+      console.error('Error deleting post.', err);
+      if (!err.statusCode) {
+        err.statusCode = 500;
+      }
+      next(err);  
+    }
+};
+
+//save post
+const savePosts = async (req, res, next) => {
+    try {
+        // Find the user
+        const user = await User.findById(req.userId);
+        if (!user) {
+            return res.status(404).json({ error: 'User not found' });
+        }
+
+        // Find the post
+        const post = await Post.findById(req.params.id);
+        if (!post) {
+            return res.status(404).json({ error: 'Post not found' });
+        }
+
+        // Check if the post is already saved by the user
+        const isSavedIndex = user.savePosts.indexOf(req.params.id);
+        const isSaved = isSavedIndex !== -1;
+
+        if (!isSaved) {
+            // Save the post
+            user.savePosts.push(req.params.id);
+            await user.save();
+            return res.status(200).json({ message: 'You saved the post.' });
+        } else {
+            // Unsave the post
+            user.savePosts.splice(isSavedIndex, 1);
+            await user.save();
+            return res.status(200).json({ message: 'You unsaved the post.' });
+        }
+    } catch (err) {
+        console.error("Error saving the post:", err);
+        return res.status(500).json({ error: 'Internal server error' });
+    }
+};
+
 
 //share post
 const  sharePost = async (req, res, next) => {
@@ -274,7 +395,7 @@ const  sharePost = async (req, res, next) => {
         }
 
         const sharedPost = new Post({
-            author: req.body.author, 
+            author: req.userId, 
             content: post.content,   
           });
       
@@ -282,10 +403,10 @@ const  sharePost = async (req, res, next) => {
           post.shares = post.shares || {}; 
           post.shares.count = post.shares.count || 0;
 
+          post.shares.users = post.shares.users || []; // Initialize users array if it doesn't exist
           post.shares.count += 1;
-          post.shares.users.push(req.body.author);
-        
-          //
+          post.shares.users.push(req.userId);
+         
           await sharedPost.save();
           await post.save();
 
@@ -305,11 +426,15 @@ module.exports = {
     addPost,
     updatePost,
     deletePost,
-    addComment,
-    deleteComment,
+    getAllPosts,
     reactPost,
+    addComment,
+    updateComment,
+    deleteComment,
     reactComment,
     replyComment,
+    reactReply,
+    deleteRely,
     savePosts,
     sharePost,
   };
